+++ db/migrate/20181115142319_create_store_item.rb
@@ -5,9 +5,9 @@ class CreateStoreItem < ActiveRecord::Migration[5.2]
       t.references :item, foreign_key: true, null: false
       t.integer :stock, null: false, default: 0
       t.integer :min_stock, null: false, default: 0
-      t.float :buy, null: false, default: 0
-      t.float :sell, null: false, default: 0
-      t.float :head_buy, null: false, default: 0
+      t.bigint :buy, null: false, default: 0
+      t.bigint :sell, null: false, default: 0
+      t.bigint :head_buy, null: false, default: 0
       
       t.timestamps
     end
