+++ db/migrate/20190319052139_create_table_capital.rb
@@ -3,7 +3,7 @@ class CreateTableCapital < ActiveRecord::Migration[5.2]
     create_table :capitals do |t|
     	t.references :store, null: false, foreign_key: true
     	t.references :user, null: false, foreign_key: true
-    	t.float :nominal, null: false
+    	t.bigint :nominal, null: false
     	t.timestamp :date_created, null: false
     	t.string :description, null: false
