+++ db/migrate/20190319065558_create_table_cash_flow.rb
@@ -3,7 +3,7 @@ class CreateTableCashFlow < ActiveRecord::Migration[5.2]
     create_table :cash_flows do |t|
     	t.references :user, foreign_key: true, null: false
     	t.references :store, foreign_key: true, null: false
-    	t.float :nominal, null: false
+    	t.bigint :nominal, null: false
     	t.integer :finance_type, null: false, default: 1
     	t.timestamp :date_created
 		  t.string :description
