+++ db/migrate/20190321080334_create_table_income.rb
@@ -3,7 +3,7 @@ class CreateTableIncome < ActiveRecord::Migration[5.2]
     create_table :incomes do |t|
     	t.references :store, null: false, foreign_key: true
     	t.references :user, null: false, foreign_key: true
-    	t.float :nominal, null: false
+    	t.bigint :nominal, null: false
     	t.timestamp :date_created, null: false
     	t.string :description, null: false
